+++ package.json
@@ -29,7 +29,9 @@
     "autoprefixer": "^10.4.21",
     "postcss": "^8.5.3",
     "react": "^19.0.0",
+    "react-cookie": "^8.0.1",
     "react-dom": "^19.0.0",
+    "react-router": "^7.6.0",
     "tailwindcss": "^4.1.6",
     "uuid": "^11.0.5",
     "zustand": "^5.0.3"
