+++ src/App.css
@@ -40,3 +40,22 @@
 .read-the-docs {
   color: #888;
 }
+
+.nav-links {
+  display: flex;
+  gap: 1rem;
+  list-style: none;
+  padding: 0;
+  justify-content: flex-end;
+}
+
+.nav-links a {
+  text-decoration: none;
+  color: gray;
+}
+
+.nav-links a.active {
+  font-weight: bold;
+  color: black;
+  border-bottom: 2px solid black;
+}
